Added option to choose to automatically delete input files

+++ atcScanner/scripts/pipeline.py
@@ -8,8 +8,6 @@ import json
 import logging
 import django
 from SNR import SNR
-import numpy as np
-import scipy.io.wavfile as wawfile
 from datetime import datetime
 
 
@@ -58,16 +56,6 @@ def get_sec(filename):
     return part[-2:]
 
 
-# def estimate_snr(audio, snr_table):
-#     # Source estimateSNR.py provided by thesis supervisor
-#     snrs = snr_table[:, 0];
-#     Gzs = snr_table[:, 1];
-#     Gz = np.log(np.mean(abs(audio) + 1.0)) - np.mean(np.log(abs(audio) + 1.0));
-#     pos = np.argmin(np.abs(Gzs - Gz));
-#     snr = snrs[pos];
-#     return snr;
-
-
 # Processes a new raw file
 def process_file(input_file, settings):
     # Change output file name to wav audio file
@@ -85,10 +73,6 @@ def process_file(input_file, settings):
         logging.error('Failed to decode file ' + output_filename.replace('.wav', settings['file_ext']))
         return
 
-    # Fs, sig = wawfile.read(output_filename)
-    # snr_table = np.loadtxt('/scripts/SNR_table_-100_100.tab')
-
-    # snr = estimate_snr(sig, snr_table)
     snr = SNR(output_filename).get_snr()
     # Delete automatically if there is too much noise
     if snr < settings['snr_thres']:
@@ -127,8 +111,14 @@ def process_file(input_file, settings):
         recording.save()
     except Exception as e:
         logging.error(f'Failed to save recording to database: {str(e)}')
-    # Delete raw file
-    os.remove(input_file)
+
+    if settings['in_autodelete']:
+        # Delete raw file
+        os.remove(input_file)
+    else:
+        # Move file to processed directory
+        os.makedirs('/data/in/processed', exist_ok=True)
+        os.replace(input_file, '/data/in/processed/' + os.path.basename(input_file))
 
 
 class GracefulShutdown:
